+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeDataEventStreams/serviceAccountsSection/serviceAccountsDataGrid/index.tsx
@@ -12,9 +12,9 @@ import { KafkaServiceAccountListDto } from '../../../../../../apis/integrations/
 import { ServiceAccountsDataGridToolbar } from './ServiceAccountsDataGridToolbar';
 import { serviceAccountsDataGridColumns } from './ServiceAccountsDataGridColumns';
 
-import { ViewCertificateDialog } from './ViewCertificateDialog';
-import { ViewKeyDialog } from './ViewKeyDialog';
-import { viewCredentialsDialogReducer } from './viewCredentialsDialogReducer';
+import { ViewCertificateDialog } from './viewCredentials/ViewCertificateDialog';
+import { ViewKeyDialog } from './viewCredentials/ViewKeyDialog';
+import { viewCredentialsDialogReducer } from './viewCredentials/viewCredentialsDialogReducer';
 
 export type ServiceAccountsDataGridProps = {
     connectionId: string;
