+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeDataEventStreams/serviceAccountsSection/serviceAccountsDataGrid/viewCredentials/viewCredentialsDialogReducer.ts
